+++ music_assistant/helpers/audio.py
@@ -63,7 +63,7 @@ if TYPE_CHECKING:
     from music_assistant.mass import MusicAssistant
     from music_assistant.models.music_provider import MusicProvider
     from music_assistant.models.player import Player
-    from music_assistant.providers.player_group import PlayerGroupProvider
+    from music_assistant.providers.universal_group import PlayerGroupProvider
 
 LOGGER = logging.getLogger(f"{MASS_LOGGER_NAME}.audio")
