--- music_assistant/providers/player_group/manifest.json
@@ -1,14 +1,0 @@
-{
-  "type": "player",
-  "domain": "player_group",
-  "stage": "beta",
-  "name": "Playergroup",
-  "description": "Create (permanent) groups of your favorite players. \nSupports both syncgroups (to group speakers of the same ecocystem to play in sync) and universal groups to group speakers of different ecosystems to play the same audio (but not in sync).",
-  "codeowners": ["@music-assistant"],
-  "requirements": [],
-  "documentation": "https://music-assistant.io/faq/groups/",
-  "multi_instance": false,
-  "builtin": true,
-  "allow_disable": false,
-  "icon": "speaker-multiple"
-}
